Временно откл.обработчики, добавляет data-index
